fixed float mismatch

+++ cot-hallucination-detection/train_value_function_with_encoder.py
@@ -363,7 +363,8 @@ class ValueModel(nn.Module):
         
         # Detach and convert to float32 for value head
         if self.freeze_base:
-            pooled = pooled.detach().float()
+            pooled = pooled.detach()
+        pooled.float()
         
         logits = self.value_head(pooled).squeeze(-1)  # (B,)
         return logits
